improve structure of extendables & user fn queueing (#1054)
* improve structure of extendables

* commit queueing work

* lint
