+++ src/lib/queueing.ts
@@ -1,0 +1,6 @@
+import PQueue from 'p-queue';
+
+/**
+ * Map<user_id, PromiseQueue>
+ */
+export const userQueues: Map<string, PQueue> = new Map();
